refactor: 將 `Entity` 更名為 `Element` 避免與原生 Entity 混淆
