--- src/source/entity/guns/Glock17.ts
@@ -1,28 +1,0 @@
-import { GunData } from "../data/Guns";
-import { Entity } from "../Entity";
-
-import { GunComponent } from "../../components/GunComponent";
-import { ItemComponent } from "../../components/ItemComponent";
-import { MagazineComponent } from "../../components/MagazineComponent";
-import { DamageComponent } from "../../components/DamageComponent";
-import { OffsetComponent } from "../../components/OffsetComponent";
-
-import { GunSystemManager } from "../../system/combat/gun/GunSystemManager";
-
-export class Glock17 extends Entity {
-
-    constructor() {
-        super('glock17');
-
-        const DATA = GunData.Glock17;
-        this.components
-            .set('item', new ItemComponent(DATA.item))
-            .set('magazine', new MagazineComponent(DATA.magazine))
-            .set('gun', new GunComponent(DATA.gun))
-            .set('offset', new OffsetComponent(DATA.offset))
-            .set('damage', new DamageComponent(DATA.damage));
-
-        GunSystemManager.instance.register(this);
-    }
-
-}
